proj3: give up on tsan
Turns out that OpenMP doesn't play nicely with ThreadSanitizer (gcc or
clang) [1][2]. Running it under tsan gives a bunch of false positives
about data races, even though everything is synchronized by the
barriers. The solution is to recompile libgomp with some options, but i
don't want to mess with that.

OpenMP doesn't play nicely with helgrind either, so valgrind is out too.
[3]

This is dumb.

[1]: https://xrunhprof.wordpress.com/2018/08/27/tsan-with-openmp/
[2]: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=55561
[3]: https://stackoverflow.com/questions/10641972/helgrind-valgrind-and-openmp-c-avoiding-false-positives#16863337