proj7b: get everything building on both flip and rabbit
The default version of gcc (4 point something) is too old to compile the
SIMD code. The newest version (8.2.0) isn't available on rabbit.
Compromise by using the latest version that rabbit has (6.3.0), which
works for both.

Note: we can't share the read.o object file with the cuda compiler,
because cuda is stuck on an old version of gcc that doesn't understand
newer object files.